+++ contracts/actions/aaveV3/helpers/AaveV3RatioHelper.sol
@@ -10,7 +10,7 @@ contract AaveV3RatioHelper is DSMath, MainnetAaveV3Addresses {
         IPoolV3 lendingPool = IPoolV3(IPoolAddressesProvider(_market).getPool());
         (uint256 totalCollUSD, uint256 totalDebtUSD, , , uint256 ltv, ) = lendingPool
             .getUserAccountData(_user);
-        if (totalDebtUSD == 0) return uint256(0);
+        if (totalDebtUSD == 0) return 0;
         return wdiv(wmul(totalCollUSD, ltv * 10**14), totalDebtUSD);
     }
 }
